Add solution for problem 13693: Cubos Coloridos
This commit adds the solution for problem 13693: Cubos Coloridos. The code implements a program that determines the number of different types of cubes in a given set. The program takes input describing the colors of the faces of each cube and outputs the number of unique types of cubes. The solution uses a combination of cube rotations and canonical forms to identify unique types of cubes.

The commit also includes the problem description and sample input/output for reference.